requirements_docs.txt
mkdocs==1.5.2
mkdocs-autorefs==0.4.1
mkdocs-click==0.8.0
mkdocs-material==9.2.7
mkdocs-redirects==1.2.1
mkdocstrings==0.22.0
mkdocstrings-python==1.1.2
pymdown-extensions==10.2.1
